Ensure sizeof(object->type) == sizeof(object->car)
There is a fundamental assumtion that object->type occupies the
same number of bytes as "object->car".  This is particularly
important for LP64, I32LP64 64-bit data models (where unsigned int
is 32-bits, and pointers are 64-bits).  Without this change,
the following code:
   (length "a").
cannot complete successfully.

+++ ulisp.c
@@ -99,7 +99,7 @@ typedef struct sobject {
       struct sobject *cdr;
     };
     struct {
-      unsigned int type;
+      uintptr_t type;
       union {
         symbol_t name;
         int integer;
